# Conflicts:
#	pom.xml
#	usl-function-web/pom.xml

+++ usl-plugin/src/main/java/com/gitee/usl/plugin/annotation/Cacheable.java
@@ -20,7 +20,7 @@ public @interface Cacheable {
      *
      * @return 数量
      */
-    long maxSize() default Long.MAX_VALUE;
+    int maxSize() default Integer.MAX_VALUE;
 
     /**
      * 缓存过期时间
